% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceback.R
\name{create_traceback}
\alias{create_traceback}
\title{Generate a traceback from a list of calls}
\usage{
create_traceback(callstack)
}
\arguments{
\item{callstack}{stack of calls, as generated by (e.g.)
\code{\link[base:sys.parent]{base::sys.calls()}}}
}
\description{
Generate a traceback from a list of calls
}
\keyword{internal}
