% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-package.R
\docType{package}
\name{evaluate-package}
\alias{evaluate-package}
\title{evaluate: Parsing and Evaluation Tools that Provide More Details than the Default}
\description{
Parsing and evaluation tools that make it easy to recreate the command line behaviour of R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://evaluate.r-lib.org/}
  \item \url{https://github.com/r-lib/evaluate}
  \item Report bugs at \url{https://github.com/r-lib/evaluate/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Yihui Xie (\href{https://orcid.org/0000-0003-0645-5666}{ORCID})
}

Other contributors:
\itemize{
  \item Michael Lawrence [contributor]
  \item Thomas Kluyver [contributor]
  \item Jeroen Ooms [contributor]
  \item Barret Schloerke [contributor]
  \item Adam Ryczkowski [contributor]
  \item Hiroaki Yutani [contributor]
  \item Michel Lang [contributor]
  \item Karolis Koncevičius [contributor]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
