\name{getSrcRegion}
\alias{getSrcRegion}
\title{Extract a rectangular region of a srcfile}
\usage{
getSrcRegion(srcfile, x1, x2, y1, y2)
}
\arguments{
  \item{srcfile}{string}

  \item{x1}{start line}

  \item{x2}{end line}

  \item{y1}{start col}

  \item{y2}{end col}
}
\value{
a string
}
\description{
Extract a rectangular region of a srcfile
}
