\name{is.message}
\alias{is.error}
\alias{is.message}
\alias{is.recordedplot}
\alias{is.source}
\alias{is.value}
\alias{is.warning}
\title{Object class tests}
\usage{
  is.error(x)

  is.message(x)

  is.source(x)

  is.value(x)

  is.warning(x)

  is.recordedplot(x)
}
\description{
  Object class tests
}
\keyword{internal}

