\name{line_prompt}
\alias{line_prompt}
\title{Line prompt.}
\usage{line_prompt(x, prompt=getOption("prompt"),
    continue=getOption("continue"))}

\description{
  Line prompt. Format a single expression as if it had been entered at
  the command prompt.
}
\value{a string}
\arguments{
  \item{x}{string representing a single expression}
  \item{prompt}{prompt for first line}
  \item{continue}{prompt for subsequent lines}
}
