\name{no_output}
\alias{no_output}
\title{Create 'no output' object}
\usage{
  no_output()
}
\value{
  an object of class \code{___no_output___}
}
\description{
  value handlers can indicate that nothing should be added
  directly to the output list by returning an object
  returned from this function
}

