% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{remove_hooks}
\alias{remove_hooks}
\title{Remove hooks.}
\usage{
remove_hooks(hooks)
}
\arguments{
\item{hooks}{a named list of hooks - each hook can either be a function or
a list of functions.}
}
\description{
This provides a way to remove previously set hook values.
}
\examples{
new1 <- list(before.plot.new = function() print("Plotted!"))
new2 <- list(before.plot.new = function() print("Plotted Again!"))
set_hooks(new1)
set_hooks(new2)
plot(1)
remove_hooks(new1)
plot(1)
remove_hooks(new2)
plot(1)
}
\keyword{internal}
