% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{replay}
\alias{replay}
\title{Replay a list of evaluated results}
\usage{
replay(x)
}
\arguments{
\item{x}{result from \code{\link[=evaluate]{evaluate()}}}
}
\description{
Replay a list of evaluated results, as if you'd run them in an R
terminal.
}
\examples{
f1 <- function() {
  cat("1\n")
  print("2")
  warning("3")
  print("4")
  message("5")
  stop("6")
}
replay(evaluate("f1()"))

f2 <- function() {
  message("Hello")
  plot(1:10)
  message("Goodbye")
}
replay(evaluate("f2()"))
}
