% Please edit documentation in R/traceback.r
\name{create_traceback}
\alias{create_traceback}
\title{Generate a traceback from a list of calls.}
\usage{
create_traceback(callstack)
}
\arguments{
\item{callstack}{stack of calls, as generated by (e.g.)
\code{\link[base]{sys.calls}}}
}
\description{
Generate a traceback from a list of calls.
}
\keyword{internal}
