% Please edit documentation in R/output.r
\name{is.message}
\alias{is.error}
\alias{is.message}
\alias{is.recordedplot}
\alias{is.source}
\alias{is.value}
\alias{is.warning}
\title{Object class tests}
\usage{
is.message(x)

is.warning(x)

is.error(x)

is.value(x)

is.source(x)

is.recordedplot(x)
}
\description{
Object class tests
}
\keyword{internal}
