\name{parse_all}
\alias{parse_all}
\title{Parse, retaining comments.}
\usage{parse_all(x)}

\description{
  Parse, retaining comments. Works very similarly to parse, but keeps
  all associated comments, so that output can be fully reconstructed.
}
\value{a data.frame}
\arguments{
  \item{x}{object to parse.  Can be a string, a file connection, or a
function}
}
