\name{watchout}
\alias{watchout}
\title{Watch for changes in output, text and graphical.}
\usage{watchout(debug=FALSE)}

\description{
  Watch for changes in output, text and graphical.
}
\value{list containing four functions: \code{get_new}, \code{pause}, 
\code{unpause}, \code{close}.}
\arguments{
  \item{debug}{activate debug mode where output will be both printed to
screen and captured.}
}
